# Improve `save_args_to_submit_buf`